Added a start for single collog syncs, Temporary fix for group collogs, have to rewrite the method to be more efficient on the db calls. Tested and succesfully written to the db.

+++ NivBot/ExternalServicesLayer/OsrsAPI/Models/Skills.cs
@@ -10,6 +10,6 @@ namespace NivBot.ExternalServicesLayer.OsrsAPI.Models
         public required string Name { get; set; }
         public required int Rank { get; set; }
         public required int Level { get; set; }
-        public required int Xp { get; set; }
+        public required long Xp { get; set; }
     }
 }
